pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Include the ABI files
[tool.setuptools.package-data]
ModelMarketSDK = ["*.json"]

[project]
name = "forumaisdk"
version = "0.0.4"
authors = [
  { name="ForumAISDK", email="forumai@github.com" },
]
description = "ForumAI Python SDK"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "tiktoken==0.6.0",
    "web3==6.15.1",
    "requests==2.31.0",
    "python-dotenv"
]

[project.urls]
Homepage = "https://github.com/0xAresDev/ForumAISDK"
Issues = "https://github.com/0xAresDev/ForumAISDK/issues"
